v0.3 tried root motion
This is where I have a weird input where E is both the roll animation and also gets the axis value, which is used to scale itself forward. Unfortunately this only works if E is held. It feels and looks really nice, but obviously we cannot make them hold the button. Currently added a root motion to new pc_roll animation that I am trying to import from blender to Unreal. (While making the skeletons and meshes play nice!)